Pridana zakladni verze MotorDriver k rizeni robota, time ve zpravach zmenen z double na long

+++ Project/src/cz/muni/ia158/PongRobot/tcp/BallInformationMessage.java
@@ -2,13 +2,13 @@ package cz.muni.ia158.PongRobot.tcp;
 
 public class BallInformationMessage {
 	private double xcoord;
-	private double time;
+	private long time;
 	private boolean isRed;
 	
 	public double getXcoord() {
 		return xcoord;
 	}
-	public double getTime() {
+	public long getTime() {
 		return time;
 	}
 	public boolean isRed() {
@@ -20,7 +20,7 @@ public class BallInformationMessage {
 	 * and when the ball will cross the "finishing line".
 	 * 
 	 */
-	public BallInformationMessage(double xcoord, double time){
+	public BallInformationMessage(double xcoord, long time){
 		this(xcoord, time, false);
 	}
 	
@@ -28,7 +28,7 @@ public class BallInformationMessage {
 	 * This constructor is for bonus assignment. If there is red ball in the field, 
 	 * then robot should care about red one, not white one.
 	 */
-	public BallInformationMessage(double xcoord, double time, boolean isRed){
+	public BallInformationMessage(double xcoord, long time, boolean isRed){
 		if(xcoord < 0){
 			throw new IllegalArgumentException("xcoord can't be negative. (or if we are using coord system in which it can be negative, than remove this line from source code of this class");
 		}
@@ -48,7 +48,7 @@ public class BallInformationMessage {
 		String params[] = p.split(";");
 		
 		double xcoord = Double.parseDouble(params[0]);
-		double time = Double.parseDouble(params[1]);
+		long time = Long.parseLong(params[1]);
 
 		boolean isRed = false;
 		if (params.length == 3){ //isRed is also there
